+++ include/axl/axl-version.h
@@ -25,9 +25,9 @@
 #define AXL_VERSION_H
 
 #define AXL_VERSION_MAJOR   0
-#define AXL_VERSION_MINOR   22
+#define AXL_VERSION_MINOR   23
 #define AXL_VERSION_PATCH   0
-#define AXL_VERSION_STRING  "0.22.0"
+#define AXL_VERSION_STRING  "0.23.0"
 
 /* Encoded as 0xMMmmpp (major, minor, patch) for simple comparisons. */
 #define AXL_VERSION_NUMBER \